requirements.txt
streamlit==1.40.2
opencv-python-headless==4.10.0.84
tensorflow==2.18.0
pillow==9.1.1
numpy==2.0.2

scikit-learn==1.3.1
pandas==2.2.3
matplotlib==3.8.0
seaborn==0.12.2
